fix: Ensure a consistent sort by assignee
This should help limit conflicts that I was seeing. Basically what
happens in that this is being run when my work computer syncs in the
other room, so the note is changed on both machines. Sometimes the sort
order wasn't identical so the note would be slightly different. This is
enough for a merge conflict, which would then propogate into other
errors.