Add rate limiter code to search_in_changed
Also updated the rate limiter values to better reflect the backoff
handler
https://github.com/laurent22/joplin/blob/6bb0318c2f9fb81302cb03985f7e59f8091f4378/packages/app-desktop/services/plugins/BackOffHandler.ts
In general, the backoff handler gives 100 requests before invoking
and if it's been less than 100 requests, it only needs a cooldown time
of 1s to reset. I've set the cooldown (scanPeriod) to 2s to add some
margin to this, and increased the count to 30 to give more room for
scanning. I still want to be a good citizen, so I didn't go to 100.