fix: Use cursor when collecting events
I don't think the page functionality actually did anything...

I also moved the note processing into the body of the get events loop
instead of collecting events. This gives slightly better memory usage in
case something happens giving us a lot of events.

This explicitly does not handle the fact that there can be multiple
events for a single note. I reason that this code is invoked after
every sync, and on each note switch. This means there is unlikely to
be any duplicates, and the code complexity will just hurt me.