Only update TODO not when opening it
Initiate sync after writing the summary
Remove the /events handling - somehow an event was missed which caused
the builder state to be wrong (and stay wrong)

These together should ensure that there are never any caused by this
plugin. It will still be possible for a user to trigger a conflict,
but it will be very rare.