Temporarily disable ical generation
This was preventing the plugin from running on mobile
(good chance it's related to this
https://github.com/adamgibbons/ics/issues/193)

I also want to somewhat conclude that experiment, it's not very elegant
having to use a proxy server, and it's not accessible to non
joplin-cloud users. I learned recently that apparently google calendar
and apple calendar support caldav, so a better calendar integration
would probably use caldav to talk to people calendars directly.