Add a sprinkling of useMemo and useCallback
I initially avoided this expecting to get away with the React compiler,
but I have lost steam and desire for using the currently experimental
esbuild plugin.
These edits focus largely on places where multiple things could
unnecessarily be re-rendered.
A remaining issue is that on many changes the global filters object will
change, which basically causes everything to re-render (e.g. checking a
single todo will trigger a re-render). There isn't a clean solution to
this that I'm happy with at the moment